make API calls to openfoodfacts API possible

+++ app/src/main/java/com/janpschwietzer/calpal/data/model/ProductModel.kt
@@ -5,7 +5,7 @@ import com.janpschwietzer.calpal.util.enums.GreenScore
 import com.janpschwietzer.calpal.util.enums.NutriScore
 
 data class ProductModel(
-    val barcode: Int,
+    val barcode: Long,
     val name: String?,
     val brand: String?,
     val kcal: Int?,
